+++ src/utils/index.js
@@ -1,7 +1,9 @@
+const ensureUpperFirst = require("./strings/ensureUpperFirst/index");
 const isLeapYear = require("./dates/isLeapYear/index.js");
 const zeroPad = require("./strings/zeroPad/index");
 
 module.exports = {
-  zeroPad,
+  ensureUpperFirst,
   isLeapYear,
+  zeroPad,
 };