+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2022-06-09
+
+### Added
+
+- Add "GET /v1/portugal/concelhos?distrito=nome"
+- Add "GET /v1/portugal/concelhos/:cod"
+- Add utils.ensureUpperFirst()
+
 ## [1.0.3] - 2022-06-09
 
 ### Added
@@ -37,4 +45,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - First release
 
 [unreleased]: https://github.com/gabtec/freeapis/compare/v1.0.0...HEAD
-[1.0.0]: https://github.com/gabtec/freeapis/releases/tag/v1.0.0
+[1.0.3]: https://github.com/gabtec/freeapis/releases/tag/v1.0.3
+[1.0.2]: https://github.com/gabtec/freeapis/releases/tag/v1.0.2
+[1.0.1]: https://github.com/gabtec/freeapis/releases/tag/v1.0.1